+++ src/Lesson5_3/L52_practice/Products/Water.java
@@ -3,7 +3,7 @@ package Lesson5_3.L52_practice.Products;
 public class Water extends Product {
     IdCeeper idCeeper;
 
-    public Water(int id, String name, int amount, String description, double price) {
+    public Water(int id, String name, int amount, String description, int price) {
         super(id, name, amount, description, price);
     }
